add the sixtysexy
The sixtysexy is a dumb name for a 66 key hybrid ergodox/atreus.
It was designed by me to fill my unique requirements. the major
feature is an additional column of the top three rows of the right
hand giving us back the [] /\ -= pairs than usually reside there
(on dvorak) removing the need to learn a new location for these keys
used quite frequenly in programming, also makes it where f12 is on
the top row (on L2). The second major feature is the size, it is designed
to fit perfectly on top of the 2015 MBP keyboard, add some wrist wrests and
you have a pretty nice setup that keeps the trackpad in the familiar place